backend/requirements.txt
openai
pydantic
jinja2
fastapi
uvicorn
python-dotenv
gunicorn